Reorganize, export react-query hooks in the package
