+++ src/__tests__/lib/containers/user_profile_links/UserProjects.test.tsx
@@ -3,15 +3,18 @@ import { render, screen } from '@testing-library/react'
 import React from 'react'
 import { mockAllIsIntersecting } from 'react-intersection-observer/test-utils'
 import UserProjects from '../../../../lib/containers/user_profile_links/UserProjects'
-import { useGetUserProjectsInfinite } from '../../../../lib/utils/hooks/SynapseAPI/useGetUserProjects'
+import { useGetUserProjectsInfinite } from '../../../../lib/utils/hooks/SynapseAPI/user/useGetUserProjects'
 import { ProjectHeader } from '../../../../lib/utils/synapseTypes'
 import { SynapseTestContext } from '../../../../mocks/MockSynapseContext'
 
-jest.mock('../../../../lib/utils/hooks/SynapseAPI/useGetUserProjects', () => {
-  return {
-    useGetUserProjectsInfinite: jest.fn(),
-  }
-})
+jest.mock(
+  '../../../../lib/utils/hooks/SynapseAPI/user/useGetUserProjects',
+  () => {
+    return {
+      useGetUserProjectsInfinite: jest.fn(),
+    }
+  },
+)
 
 const mockFetchNextPage = jest.fn()
 const mockUseGetUserProjectsInfinite = useGetUserProjectsInfinite as jest.Mock
