+++ src/lib/containers/SynapseNavDrawer.tsx
@@ -5,8 +5,8 @@ import SynapseIconWhite from '../assets/icons/SynapseIconWhite'
 import SynapseLogoName from '../assets/icons/SynapseLogoName'
 import { SynapseClient } from '../utils'
 import { useSearchAccessSubmissionsInfinite } from '../utils/hooks/SynapseAPI/dataaccess/useDataAccessSubmission'
-import { useGetDownloadListStatistics } from '../utils/hooks/SynapseAPI/useGetDownloadListStatistics'
-import { useGetCurrentUserBundle } from '../utils/hooks/SynapseAPI/useUserBundle'
+import { useGetDownloadListStatistics } from '../utils/hooks/SynapseAPI/download/useGetDownloadListStatistics'
+import { useGetCurrentUserBundle } from '../utils/hooks/SynapseAPI/user/useUserBundle'
 import { isInSynapseExperimentalMode } from '../utils/SynapseClient'
 import { useSynapseContext } from '../utils/SynapseContext'
 import { Direction, SubmissionState } from '../utils/synapseTypes'
