+++ src/lib/containers/UserCardMedium.tsx
@@ -6,7 +6,7 @@ import ValidatedProfileIcon from '../assets/icons/ValidatedProfile'
 import { SkeletonTable } from '../assets/skeletons/SkeletonTable'
 import { SynapseConstants } from '../utils'
 import { PRODUCTION_ENDPOINT_CONFIG } from '../utils/functions/getEndpoint'
-import { useGetUserBundle } from '../utils/hooks/SynapseAPI/useUserBundle'
+import { useGetUserBundle } from '../utils/hooks/SynapseAPI/user/useUserBundle'
 import { UserProfile } from '../utils/synapseTypes/'
 import { Avatar } from './Avatar'
 import IconSvg from './IconSvg'
