+++ src/lib/containers/download_list/DownloadConfirmation.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react'
 import { Alert } from 'react-bootstrap'
 import useDeepCompareEffect from 'use-deep-compare-effect'
 import { SynapseClient, SynapseConstants } from '../../utils'
-import { useGetDownloadListStatistics } from '../../utils/hooks/SynapseAPI/useGetDownloadListStatistics'
+import { useGetDownloadListStatistics } from '../../utils/hooks/SynapseAPI/download/useGetDownloadListStatistics'
 import { useGetEntityChildren } from '../../utils/hooks/SynapseAPI/entity/useGetEntityChildren'
 import useGetQueryResultBundle from '../../utils/hooks/SynapseAPI/entity/useGetQueryResultBundle'
 import { useSynapseContext } from '../../utils/SynapseContext'
