+++ src/lib/containers/download_list_v2/DownloadListActionsRequired.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react'
 import { useErrorHandler } from 'react-error-boundary'
-import { useGetDownloadListActionsRequiredInfinite } from '../../utils/hooks/SynapseAPI/useGetDownloadListActionsRequired'
+import { useGetDownloadListActionsRequiredInfinite } from '../../utils/hooks/SynapseAPI/download/useGetDownloadListActionsRequired'
 import { useInView } from 'react-intersection-observer'
 import {
   ActionRequiredCount,
