+++ src/lib/containers/entity/annotations/SchemaDrivenAnnotationEditor.tsx
@@ -14,7 +14,7 @@ import {
   useUpdateViaJson,
 } from '../../../utils/hooks/SynapseAPI/entity/useEntity'
 import { useGetSchemaBinding } from '../../../utils/hooks/SynapseAPI/entity/useEntityBoundSchema'
-import { useGetSchema } from '../../../utils/hooks/SynapseAPI/useSchema'
+import { useGetSchema } from '../../../utils/hooks/SynapseAPI/entity/useSchema'
 import { SynapseClientError } from '../../../utils/SynapseClientError'
 import { EntityJson, entityJsonKeys } from '../../../utils/synapseTypes'
 import { SynapseSpinner } from '../../LoadingScreen'
