+++ src/lib/containers/user_profile_links/UserProjects.tsx
@@ -3,7 +3,7 @@ import { useErrorHandler } from 'react-error-boundary'
 import { useInView } from 'react-intersection-observer'
 import { ProjectHeader } from '../../utils/synapseTypes'
 import { PRODUCTION_ENDPOINT_CONFIG } from '../../utils/functions/getEndpoint'
-import { useGetUserProjectsInfinite } from '../../utils/hooks/SynapseAPI/useGetUserProjects'
+import { useGetUserProjectsInfinite } from '../../utils/hooks/SynapseAPI/user/useGetUserProjects'
 import { GetProjectsParameters } from '../../utils/synapseTypes/GetProjectsParams'
 import { SkeletonTable } from '../../assets/skeletons/SkeletonTable'
