+++ src/lib/utils/hooks/SynapseAPI/dataaccess/useGetApprovedSubmissionInfo.ts
@@ -1,11 +1,11 @@
 import { useInfiniteQuery, UseInfiniteQueryOptions } from 'react-query'
-import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClientError'
-import { useSynapseContext } from '../../SynapseContext'
+import { SynapseClient } from '../../..'
+import { SynapseClientError } from '../../../SynapseClientError'
+import { useSynapseContext } from '../../../SynapseContext'
 import {
   SubmissionInfoPage,
   SubmissionInfoPageRequest,
-} from '../../synapseTypes/SubmissionInfo'
+} from '../../../synapseTypes/SubmissionInfo'
 
 export function useGetApprovedSubmissionInfoInfinite(
   accessRequirementId: string,
