+++ src/lib/utils/hooks/SynapseAPI/entity/useSchema.ts
@@ -4,8 +4,8 @@
 
 import { JSONSchema7 } from 'json-schema'
 import { useQuery, UseQueryOptions } from 'react-query'
-import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClientError'
+import { SynapseClient } from '../../..'
+import { SynapseClientError } from '../../../SynapseClientError'
 
 export function useGetSchema(
   schema$id: string,
