+++ src/lib/utils/hooks/SynapseAPI/search/index.ts
@@ -1,0 +1,1 @@
+export * from './useSearch'
