+++ src/lib/utils/hooks/SynapseAPI/user/useFavorites.ts
@@ -4,10 +4,10 @@ import {
   UseQueryOptions,
   useInfiniteQuery,
 } from 'react-query'
-import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClientError'
-import { useSynapseContext } from '../../SynapseContext'
-import { EntityHeader, PaginatedResults } from '../../synapseTypes'
+import { SynapseClient } from '../../..'
+import { SynapseClientError } from '../../../SynapseClientError'
+import { useSynapseContext } from '../../../SynapseContext'
+import { EntityHeader, PaginatedResults } from '../../../synapseTypes'
 
 export function useGetFavorites(
   options?: UseQueryOptions<
