+++ src/lib/utils/hooks/SynapseAPI/user/useGetUserChallenges.ts
@@ -1,8 +1,8 @@
 import { useInfiniteQuery, UseInfiniteQueryOptions } from 'react-query'
-import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClientError'
-import { useSynapseContext } from '../../SynapseContext'
-import { ChallengeWithProjectHeaderPagedResults } from '../../synapseTypes/ChallengePagedResults'
+import { SynapseClient } from '../../..'
+import { SynapseClientError } from '../../../SynapseClientError'
+import { useSynapseContext } from '../../../SynapseContext'
+import { ChallengeWithProjectHeaderPagedResults } from '../../../synapseTypes/ChallengePagedResults'
 
 export function useGetUserChallengesInfinite(
   userId: string,
