+++ src/lib/utils/hooks/SynapseAPI/user/useGetUserProjects.ts
@@ -1,9 +1,9 @@
 import { useInfiniteQuery, UseInfiniteQueryOptions } from 'react-query'
-import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClientError'
-import { useSynapseContext } from '../../SynapseContext'
-import { ProjectHeaderList } from '../../synapseTypes'
-import { GetProjectsParameters } from '../../synapseTypes/GetProjectsParams'
+import { SynapseClient } from '../../..'
+import { SynapseClientError } from '../../../SynapseClientError'
+import { useSynapseContext } from '../../../SynapseContext'
+import { ProjectHeaderList } from '../../../synapseTypes'
+import { GetProjectsParameters } from '../../../synapseTypes/GetProjectsParams'
 
 export function useGetUserProjectsInfinite(
   userId: string,
