+++ src/lib/utils/hooks/SynapseAPI/user/useProjects.ts
@@ -4,11 +4,11 @@ import {
   useQuery,
   UseQueryOptions,
 } from 'react-query'
-import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClientError'
-import { useSynapseContext } from '../../SynapseContext'
-import { ProjectHeaderList } from '../../synapseTypes'
-import { GetProjectsParameters } from '../../synapseTypes/GetProjectsParams'
+import { SynapseClient } from '../../..'
+import { SynapseClientError } from '../../../SynapseClientError'
+import { useSynapseContext } from '../../../SynapseContext'
+import { ProjectHeaderList } from '../../../synapseTypes'
+import { GetProjectsParameters } from '../../../synapseTypes/GetProjectsParams'
 
 export function useGetProjects(
   params?: GetProjectsParameters,
