+++ src/lib/utils/index.tsx
@@ -1,6 +1,7 @@
 import * as SynapseConstants from './SynapseConstants'
 import * as SynapseClient from './SynapseClient'
+import * as SynapseQueries from './hooks/SynapseAPI'
 import { RequiredProperties } from './types/RequiredProperties'
 
-export { SynapseConstants, SynapseClient }
+export { SynapseConstants, SynapseClient, SynapseQueries }
 export type { RequiredProperties }
